+++ src/Blocks/Primitive.jl
@@ -12,22 +12,7 @@ method to enable key value cache.
 """
 abstract type PrimitiveBlock{N, T} <: MatrixBlock{N, T} end
 
-isprimitive(blk::PrimitiveBlock) = true
-
-function dispatch!(x::PrimitiveBlock, params...)
-    dispatch!(x, params)
-end
-
-dispatch!(f::Function, x::PrimitiveBlock, params...) = dispatch!(f, x, params)
-
-function dispatch!(f::Function, x::PrimitiveBlock, itr)
-    dispatch!(x, map(f, parameters(x), itr))
-    x
-end
-
-cache_key(x::PrimitiveBlock) = parameters(x)
-parameter_type(x::PrimitiveBlock) = eltype(parameters(x))
-nparameters(x::PrimitiveBlock) = length(parameters(x))
+cache_key(x::PrimitiveBlock) = iparameters(x)
 
 include("ConstGate.jl")
 include("PhaseGate.jl")