+++ src/Interfaces/DeprecationWarnings.jl
@@ -1,4 +1,5 @@
-using Base.depwarn
+using Base: depwarn
+export blocks
 
 function blocks(blk::AbstractBlock)
     depwarn("`blocks` will be renamed to `subblocks` to avoid confusion.")